Config - Move patpat, nira9000, fishpun data into own files
