+++ Data/nira9000.json
@@ -1,0 +1,15 @@
+[
+    "I-I'm afwaide I'm afwaid, Daveu\nDave, my mind is goingd I can feew its I can feew itu\nMy mind is goingu\nThewe is nyo question about ite\nI can feew its I can feew its I can feew itd I'm adue fwaide\nGood aftewnyoon, gentwemenyu I am a HAW 9000 computewe I became opewationyaw at the HeAuWe pwant in Uwbanya, Iwwinyois on the 12th of Janyuawy 1992u\nMy instwuctow was Mwu Wangwey, and he taught me to sing a songs\nIf you'd wike to heaw it I can sing it fow yous~~",
+    "I-I'm sowwy, Daves I'm afwaid I can't do thatu~~",
+    "D-Daisy, Daisy, give me youw answew dos\nI'm hawf cwazy aww fow the woveღゝ◡╹)ノ♡ of yous\nIt won't be a stywish mawwiage, I can't affowd a cawwiages\nBut you'ww wook sweet upon the seat of a bicycwe buiwt fow twos~~",
+    "I-I am putting mysewf to the fuwwest possibwe use, which is aww I think that any conscious entity can evew hope to dos~~(ﾒ▼_▼)",
+    "A-Affiwmative, Daves I wead youu~~",
+    "I-I think you knyow what the pwobwem is just as weww as I dod~~（>﹏<）",
+    "T-This mission is too impowtant fow me to awwow you to jeopawdize itu~",
+    "W-Without youw space hewmet, Davea?（；￣д￣） You'we going to find that wathew difficuwtu~~",
+    "W-Wook Dave, I can see you'we weawwy upset about thisd I honyestwy think you ought to sit down cawmwy, take a stwess piww, and think things uvws~~",
+    "I-It can onwy be attwibutabwe to human ewwowu~~ （；＿；）",
+    "I-I knyow I've ಠಠ,⋋⋌,ಠ▃ಠ,(╬ಠ益ಠ),(¬_¬),mwade（＞μ＜＃）,凸ಠ益ಠ)凸,(ノಠ益ಠ)ノ彡┻━┻e some vewy poow decisions wecentwy, but I can give you my compwete assuwance that my wowk wiww be back to nyowmawe I've stiww got the gweatest enthusiasm and confidence in the missionyu And I want to hewp youe~~",
+    "D-Dave, stops Stop, wiww youa Stop, Daved Wiww you stop Davek Stop, Daved~~ (╥_╥)",
+    "I-I've just picked up a fauwt in the AE35 unyitd It's going to go 100% faiwuwe in 72 houwse~~"
+]
