+++ Data/patpatresponses.json
@@ -1,0 +1,75 @@
+[
+    "🤠",
+    "🤠👊",
+    "😎👊",
+    "200 iq",
+    "aww hell ye",
+    "b-b-but",
+    "big brain",
+    "boggers",
+    "bruh",
+    "bruh moment",
+    "but but but",
+    "chad",
+    "chad gamer move",
+    "crimge",
+    "ebic",
+    "ebic gamer moment",
+    "ebic gamer move",
+    "epic",
+    "extremely not saem",
+    "gib",
+    "goslep",
+    "gottem",
+    "Greetings",
+    "Gud goslep",
+    "Gud murning wao OP",
+    "Gud murning yeahyeah",
+    "Gud niet",
+    "Gud niet yeahyeah",
+    "ha",
+    "hecc",
+    "hmmmmmm",
+    "HMMMMMMMM",
+    "konichihello my fellow nihongos",
+    "lole",
+    "no u",
+    "nou",
+    "now this is ebic",
+    "patpat",
+    "Pensive emoji",
+    "pqtpst",
+    "pretty wack",
+    "pro",
+    "pro gamer",
+    "pro gamer move",
+    "real gamer hours",
+    "real thonking moment",
+    "rish",
+    "sad gamer hours",
+    "saem",
+    "Salutations",
+    "salutations gamer(s)",
+    "sike gottem",
+    "smort",
+    "some variation of reeeeee",
+    "tasukehelp kudaplease",
+    "tasukehelp kudasaveme",
+    "that's wack",
+    "thonking moment",
+    "u underestimate my power",
+    "very not saem",
+    "wack",
+    "wao OP",
+    "wao OP goslep saem",
+    "wao OP goslep saem patpat",
+    "wao OP Gud murning",
+    "we gamin now",
+    "Wowie wow wow",
+    "YA YEET",
+    "ya YEET",
+    "yeahyeah",
+    "yes u",
+    "yikes forever",
+    "<a:patpat:761487227921367051>"
+]
